refactor about page to use global context from layout
